requirements.txt
fastapi
uvicorn[standard]
pymongo
dnspython
google-genai
python-dotenv
